+++ packages/whisky-common/Cargo.toml
@@ -6,6 +6,7 @@ license = "Apache-2.0"
 description = "The Cardano Rust SDK, inspired by MeshJS"
 
 [dependencies]
+uplc = "=1.1.9"
 hex = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
