+++ packages/whisky-common/src/interfaces/fetcher.rs
@@ -1,20 +1,7 @@
+use std::collections::HashMap;
+
+use crate::*;
 use async_trait::async_trait;
-use serde::{Deserialize, Serialize};
-
-use crate::errors::*;
-use crate::models::*;
-
-pub struct FetcherOptions {
-    pub max_page: Option<usize>,
-    pub order: Option<FetchOrder>,
-    pub additional_options: std::collections::HashMap<String, serde_json::Value>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub enum FetchOrder {
-    Asc,
-    Desc,
-}
 
 #[async_trait]
 pub trait Fetcher: Send + Sync {
@@ -24,26 +11,20 @@ pub trait Fetcher: Send + Sync {
         address: &str,
         asset: Option<&str>,
     ) -> Result<Vec<UTxO>, WError>;
-    async fn fetch_address_txs(
-        &self,
-        address: &str,
-        options: Option<FetcherOptions>,
-    ) -> Result<Vec<TransactionInfo>, WError>;
+
     async fn fetch_asset_addresses(&self, asset: &str) -> Result<Vec<(String, String)>, WError>;
-    async fn fetch_asset_metadata(&self, asset: &str) -> Result<AssetMetadata, WError>;
+    async fn fetch_asset_metadata(
+        &self,
+        asset: &str,
+    ) -> Result<Option<HashMap<String, serde_json::Value>>, WError>;
     async fn fetch_block_info(&self, hash: &str) -> Result<BlockInfo, WError>;
     async fn fetch_collection_assets(
         &self,
         policy_id: &str,
         cursor: Option<String>,
-    ) -> Result<(Vec<(String, String)>, Option<(u64, String)>), WError>;
-    async fn fetch_protocol_parameters(&self, epoch: u32) -> Result<Protocol, WError>;
+    ) -> Result<(Vec<(String, String)>, Option<String>), WError>;
+    async fn fetch_protocol_parameters(&self, epoch: Option<u32>) -> Result<Protocol, WError>;
     async fn fetch_tx_info(&self, hash: &str) -> Result<TransactionInfo, WError>;
-    async fn fetch_utxos(&self, hash: &str, index: Option<usize>) -> Result<Vec<UTxO>, WError>;
-    async fn fetch_governance_proposal(
-        &self,
-        tx_hash: &str,
-        cert_index: usize,
-    ) -> Result<GovernanceProposalInfo, WError>;
+    async fn fetch_utxos(&self, hash: &str, index: Option<u32>) -> Result<Vec<UTxO>, WError>;
     async fn get(&self, url: &str) -> Result<serde_json::Value, WError>;
 }
