+++ packages/whisky-common/src/interfaces/mod.rs
@@ -1,3 +1,4 @@
+mod evaluator;
 mod fetcher;
 mod provider;
 mod serializer;
@@ -5,6 +6,7 @@ mod tx_builder;
 mod tx_parser;
 // pub use provider::*;
 // pub use serializer::*;
+pub use evaluator::*;
 pub use fetcher::*;
 pub use tx_builder::*;
 // pub use tx_parser::*;
