+++ packages/whisky-csl/Cargo.toml
@@ -21,8 +21,6 @@ whisky-common = { version = "1.0.0-alpha.1", path = "../whisky-common" }
 
 # non-wasm
 [target.'cfg(not(all(target_arch = "wasm32", not(target_os = "emscripten"))))'.dependencies]
-# uplc = { version = "=1.0.26-alpha", default-features = false, features = ["native-secp256k1"] }
-# uplc = "=1.0.26-alpha"
 uplc = "=1.1.9"
 wasm-bindgen = { version = "=0.2.92", features = ["serde-serialize"] }
 rand_os = "0.1"
@@ -35,7 +33,6 @@ tiny-bip39 = "2.0.0"
 # wasm
 [target.'cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))'.dependencies]
 serde-wasm-bindgen = "0.4.5"
-# uplc = { version = "=1.0.26-alpha", default-features = false, features = ["native-secp256k1"] }
 uplc = "=1.1.9"
 wasm-bindgen = { version = "=0.2.92", features = ["serde-serialize"] }
 rand_os = { version = "0.1", features = ["wasm-bindgen"] }
