+++ packages/whisky/src/providers/maestro/models/asset.rs
