+++ packages/whisky/src/services/fetcher.rs
@@ -1,0 +1,1 @@
+
