+++ packages/whisky/src/services/mod.rs
@@ -3,5 +3,5 @@ mod fetcher;
 mod submitter;
 
 pub use evaluator::*;
-pub use fetcher::*;
+// pub use fetcher::*;
 pub use submitter::*;
